requirements.txt
numpy==1.26.4
pytorch_msssim==0.2.1
pytz==2022.5
torchsummary==1.5.1
timm==1.0.7
torch==2.1.0
torchvision==0.16.0
torchmetrics==1.1.2
pytorch-lightning==2.4.0
lightly==1.5.13
tensorboard==2.18.0
rasterio==1.4.2
pillow==11.0.0
huggingface-hub==0.26.2
transformers==4.39.0
pandas==2.2.3
tqdm==4.65.1

